:bug: fix-be: SinglePlay 점수 계산 로직 변경

See merge request s10-bigdata-recom-sub2/S10P22A106!401

+++ englising-be/src/main/java/org/englising/com/englisingbe/singleplay/service/SinglePlayServiceImpl.java
@@ -214,7 +214,7 @@ public class SinglePlayServiceImpl {
                 return 0;
             }
             int total = 30+40+50;
-            int rate = totalScore / total * 100;
+            double rate = (double)totalScore / total * 100;
             if (rate < 30) {
                 score = 0;
             } else if (rate < 60) {
